Add CI for running tests on pushes (#6)
* ci: add workflow for regression

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

* ci: fix

+++ switch/testbench/utility.h
@@ -8,8 +8,8 @@
 
 template <typename T>
 int ensure(T actual, const std::span<T> &expected, const char *test_name) {
-    extern int fails;
-    extern int sim_time;
+    extern uint64_t fails;
+    extern uint64_t sim_time;
     bool found = false;
     int i = 0;
     for (i = 0; i < expected.size(); i++) {
